Re-organize file structure. Add new layout.jsx file
