+++ app/(with-homepage-header)/page.js
@@ -1,4 +1,4 @@
-import Homepage from "./Homepage";
+import Homepage from "../Homepage";
 
 export default function Home() {
   return (
