+++ app/(with-homepage-header)/reset-password/[token]/page.jsx
@@ -1,4 +1,3 @@
-
 const ResetPassword = async ({ params }) => {
 
   const { token } = await params;
