--- app/components/HomepageHeader/CreateAccountButton.jsx
@@ -1,15 +1,0 @@
-'use client'
-
-import Button from 'react-bootstrap/Button'
-
-function CreateAccountButton(props) {
-    return (
-      <>
-        <Button className="ms-auto" variant="outline-primary" onClick={props.openModal}>
-          Create Account
-        </Button>        
-      </>
-    );
-}
-
-export default CreateAccountButton;
