--- app/components/HomepageHeader/LogInButton.jsx
@@ -1,15 +1,0 @@
-'use client'
-
-import Button from 'react-bootstrap/Button'
-
-function LogInButton(props) {
-    return (
-      <>
-        <Button className="ms-auto" variant="primary" onClick={props.openModal}>
-          Log In
-        </Button>
-      </>
-    );
-}
-
-export default LogInButton;
