+++ app/components/headers/ProfilePictureToggler.jsx
